feat:creation login avec generation token

+++ database/migrations/2026_07_28_123434_create_role_permissions_table.php
@@ -8,14 +8,16 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('role_permissions', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('role_id')->constrained('roles')->onDelete('cascade');  // <-- CORRIGÉ
-            $table->foreignId('permission_id')->constrained('permissions')->onDelete('cascade');
-            $table->timestamps();
+        if (!Schema::hasTable('role_permission')) {
+            Schema::create('role_permission', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('role_id')->constrained('roles')->onDelete('cascade');
+                $table->foreignId('permission_id')->constrained('permissions')->onDelete('cascade');
+                $table->timestamps();
 
-            $table->unique(['role_id', 'permission_id']);
-        });
+                $table->unique(['role_id', 'permission_id']);
+            });
+        }
     }
 
     public function down(): void
